requirements.txt
httpx
asyncio
beautifulsoup4
pycryptodome
